[refactor] Package Moving Dibs and Feeling #166
