+++ app/src/main/java/find/ui/ui/dibs/DibsAdapter.kt
@@ -1,9 +1,12 @@
-package find.ui.ui.matching
+package find.ui.ui.dibs
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import find.ui.databinding.ItemDibsBinding
+import find.ui.ui.matching.MatchingDiffUtil
+import find.ui.ui.matching.MatchingViewHolder
+import find.ui.ui.matching.ResponseMatching
 
 class DibsAdapter(private val isMyDibs: Boolean) :
     ListAdapter<ResponseMatching, MatchingViewHolder<ItemDibsBinding>>(MatchingDiffUtil()) {
