+++ app/src/main/java/find/ui/ui/feeling/FeelingAdapter.kt
@@ -1,15 +1,18 @@
-package find.ui.ui.matching
+package find.ui.ui.feeling
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import find.ui.databinding.ItemFeelingSendGetBinding
+import find.ui.ui.matching.MatchingDiffUtil
+import find.ui.ui.matching.MatchingViewHolder
+import find.ui.ui.matching.ResponseMatching
 
 class FeelingAdapter(private val isSend: Boolean) :
     ListAdapter<ResponseMatching, MatchingViewHolder<ItemFeelingSendGetBinding>>(MatchingDiffUtil()) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) =
-        MatchingViewHolder<ItemFeelingSendGetBinding>(
+        MatchingViewHolder(
             ItemFeelingSendGetBinding.inflate(
                 LayoutInflater.from(parent.context),
                 parent,
