+++ app/src/main/java/find/ui/ui/feeling/FeelingConnectedAdapter.kt
@@ -1,9 +1,12 @@
-package find.ui.ui.matching
+package find.ui.ui.feeling
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import find.ui.databinding.ItemFeelingConnectedBinding
+import find.ui.ui.matching.MatchingDiffUtil
+import find.ui.ui.matching.MatchingViewHolder
+import find.ui.ui.matching.ResponseMatching
 
 class FeelingConnectedAdapter :
     ListAdapter<ResponseMatching, MatchingViewHolder<ItemFeelingConnectedBinding>>(
@@ -11,7 +14,7 @@ class FeelingConnectedAdapter :
     ) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) =
-        MatchingViewHolder<ItemFeelingConnectedBinding>(
+        MatchingViewHolder(
             ItemFeelingConnectedBinding.inflate(
                 LayoutInflater.from(parent.context),
                 parent,
